+++ .idea/compiler.xml
@@ -6,14 +6,19 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="Factions Module" />
+        <module name="NexusFactionsAddon" />
         <module name="NexusCore" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
       <module name="BRAWLMC-NEXUS" target="10" />
+      <module name="Factions Module" target="1.8" />
       <module name="NexusAPI" target="17" />
       <module name="NexusCore" target="1.8" />
+      <module name="NexusCore2" target="1.8" />
       <module name="NexusFactions" target="10" />
+      <module name="NexusFactionsAddon" target="1.8" />
     </bytecodeTargetLevel>
   </component>
 </project>
