+++ .idea/encodings.xml
@@ -3,7 +3,7 @@
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/NexusCore/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/NexusCore/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/NexusFactions/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/NexusFactions/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/NexusFactionsAddon/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/NexusFactionsAddon/src/main/resources" charset="UTF-8" />
   </component>
 </project>
