--- .idea/libraries/Maven__com_massivecraft_Factions_Plugin_1_6_9_5_U0_6_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.massivecraft:Factions-Plugin:1.6.9.5-U0.6.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/massivecraft/Factions-Plugin/1.6.9.5-U0.6.0/Factions-Plugin-1.6.9.5-U0.6.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/massivecraft/Factions-Plugin/1.6.9.5-U0.6.0/Factions-Plugin-1.6.9.5-U0.6.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/massivecraft/Factions-Plugin/1.6.9.5-U0.6.0/Factions-Plugin-1.6.9.5-U0.6.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
