--- .idea/libraries/Maven__net_kyori_adventure_platform_facet_4_0_0_SNAPSHOT.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.kyori:adventure-platform-facet:4.0.0-SNAPSHOT">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-platform-facet/4.0.0-SNAPSHOT/adventure-platform-facet-4.0.0-20210914.041527-54.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-platform-facet/4.0.0-SNAPSHOT/adventure-platform-facet-4.0.0-20210914.041527-54-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-platform-facet/4.0.0-SNAPSHOT/adventure-platform-facet-4.0.0-20210914.041527-54-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
