--- .idea/libraries/Maven__net_kyori_adventure_text_minimessage_4_2_0_SNAPSHOT.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.kyori:adventure-text-minimessage:4.2.0-SNAPSHOT">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-text-minimessage/4.2.0-SNAPSHOT/adventure-text-minimessage-4.2.0-20211031.053702-35.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-text-minimessage/4.2.0-SNAPSHOT/adventure-text-minimessage-4.2.0-20211031.053702-35-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-text-minimessage/4.2.0-SNAPSHOT/adventure-text-minimessage-4.2.0-20211031.053702-35-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
