--- .idea/libraries/Maven__net_kyori_adventure_text_serializer_gson_legacy_impl_4_9_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: net.kyori:adventure-text-serializer-gson-legacy-impl:4.9.1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-text-serializer-gson-legacy-impl/4.9.1/adventure-text-serializer-gson-legacy-impl-4.9.1.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-text-serializer-gson-legacy-impl/4.9.1/adventure-text-serializer-gson-legacy-impl-4.9.1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/net/kyori/adventure-text-serializer-gson-legacy-impl/4.9.1/adventure-text-serializer-gson-legacy-impl-4.9.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
