+++ .idea/misc.xml
@@ -10,7 +10,7 @@
       <list>
         <option value="$PROJECT_DIR$/pom.xml" />
         <option value="$PROJECT_DIR$/NexusCore/pom.xml" />
-        <option value="$PROJECT_DIR$/NexusFactions/pom.xml" />
+        <option value="$PROJECT_DIR$/NexusFactionsAddon/pom.xml" />
       </list>
     </option>
   </component>
