+++ .idea/modules.xml
@@ -3,7 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/NexusCore/NexusCore.iml" filepath="$PROJECT_DIR$/NexusCore/NexusCore.iml" />
-      <module fileurl="file://$PROJECT_DIR$/NexusFactions/NexusFactions.iml" filepath="$PROJECT_DIR$/NexusFactions/NexusFactions.iml" />
+      <module fileurl="file://$PROJECT_DIR$/NexusFactionsAddon/NexusFactionsAddon.iml" filepath="$PROJECT_DIR$/NexusFactionsAddon/NexusFactionsAddon.iml" />
     </modules>
   </component>
 </project>
