--- NexusFactions/src/main/resources/plugin.yml
@@ -1,11 +1,0 @@
-name: FactionsNexus
-version: '${project.version}'
-main: brawl.factionsnexus.FactionsNexus
-api-version: 1.17
-prefix: Faction Nexus
-depend:
-  - "Factions"
-  - "NexusCore"
-authors: [ millemila ]
-description: Faction nexus to protect factions from other players
-website: brawlmc.com
