+++ NexusFactionsAddon/src/main/java/brawl/factionsnexus/FactionsNexus.java
@@ -1,0 +1,69 @@
+package brawl.factionsnexus;
+
+import com.massivecraft.factions.Board;
+import com.massivecraft.factions.Factions;
+import listeners.*;
+import org.bukkit.plugin.PluginManager;
+import org.bukkit.plugin.java.JavaPlugin;
+import util.FactionOperations;
+
+
+public final class FactionsNexus extends JavaPlugin {
+
+    FactionsNexusController nexusController;
+    FactionOperations nexusOperations;
+
+    @Override
+    public void onEnable() {
+
+
+
+        nexusController = new FactionsNexusController(this, getFactionsInstance(), getBoardInstance());
+
+        writeDefaultConfig();
+        registerListeners();
+
+    }
+
+    Board getBoardInstance(){
+        Board board = Board.getInstance();
+        if (board == null){
+            return null;
+        }
+        return board;
+    }
+
+    Factions getFactionsInstance(){
+        Factions factions = Factions.getInstance();
+        if (factions == null){
+            return null;
+        }
+        return factions;
+    }
+
+    @Override
+    public void onDisable() {
+
+    }
+
+    private void writeDefaultConfig()
+    {
+        this.saveDefaultConfig();
+    }
+
+    private void registerListeners()
+    {
+        PluginManager pluginManager = getServer().getPluginManager();
+
+        pluginManager.registerEvents(new FactionCreateListener(), this);
+        pluginManager.registerEvents(new FactionSetHomeListener(), this);
+        pluginManager.registerEvents(new FactionLeaveListener(),this);
+        pluginManager.registerEvents(new FactionDisbandListener(), this);
+        pluginManager.registerEvents(new BlockPlaceListener(), this);
+        pluginManager.registerEvents(new BlockBreakListener(),this);
+        pluginManager.registerEvents(new FactionUnclaimListener(),this);
+        pluginManager.registerEvents(new FactionUnclaimAllListener(),this);
+
+    }
+
+}
