+++ NexusFactionsAddon/src/main/resources/plugin.yml
@@ -1,0 +1,9 @@
+name: FactionsModule
+version: '${project.version}'
+main: brawl.factionsmodule.FactionsModule
+api-version: 1.17
+prefix: Nexus Factions Add-on
+depend: [NexusCore]
+authors: [millemila]
+description: Factions add-on for NexusCore
+website: brawlmc.com
